-fix nomenclatura,csvVisualizador pra csvLikedin
